[OGUI-723] Improve 404 message consul for links configuration page (#659)
By default, Consul will respond with 404 if no key is found with the requested prefix. Due to this, an error message will be shown to the user that there was a problem with Consul.

This change will improve the message on the links configuration page so that if Consul replies with 404, a custom error message will be displayed informing the user about the keys that we tried to retrieve information for and the fact that there were none.

By doing so, the user will be able to check themselves if the FLP deployment was correct and why the configurations are not stored as expected.

Moreover, the sizes of the error and loading icon pages were reduced.