[OGUI-1722] Runs mode layout smart refresh (#3078)
This PR removes the unnecessary objects refresh in layout view by checking their ids and comparing them with the previous one loaded (in run mode) when a run is ongoing. It also removes the property of run status of being RemoteData, so that we avoid an unnecessary blink when it changes.