[OGUI-1710] Handle newly generated token from central system (#2992)
In the scope of this PR I have created new command for token creation. For now, ConnectionDirection.DUPLEX is handled like two different Connection instances for Sending and Receiving. If there will be a need for duplex connection we can change that. I believe there is no need for that, besides the one we have for CentralConnection.